Vercel/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.69
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==0.4.3
numpy==1.24.1
openai==0.26.2
pathspec==0.10.3
platformdirs==2.6.2
pydantic==1.10.4
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.2
SQLAlchemy==1.4.46
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
